+++ Services/User/IUserService.cs
@@ -11,8 +11,8 @@ namespace Services.User {
         public ReauthenticateResponse ReauthenticateUser(ReauthenticateRequest data);
         #nullable enable
         public Domain.User? GetUser(long id);
-        public string SaveUserProfileImage(long id, byte[] image);
-        public void ChangeUserAvatarPath(long id, string path);
+        public string SaveUserProfileImage(int id, byte[] image);
+        public void ChangeUserAvatarPath(int id, string path);
         public void UpdateUser(Domain.User user);
         public bool IsUserExists(long id);
